requirements.txt
torch
numpy
scipy
numba
scikit-learn
g2p_en
edit_distance
hydra-core
transformers